Keys larger than 32 bits are now able to be handled.

+++ server/src/test/java/dev/findfirst/CryptoProjectFinal/CryptoProjectFinalApplicationTests.java
@@ -36,7 +36,7 @@ class CryptoProjectFinalApplicationTests {
 
   @Test
   void findPrimeForBitSize() {
-    int priKey = keyGen.genRandomInBitSize(16);
+    long priKey = keyGen.genRandomInBitSize(16);
     while (!keyGen.primeTest(priKey, 4)) {
       priKey = keyGen.genRandomInBitSize(16);
     }
